cleaning up DeviceTypes import; test fixes
