+++ src/cryptoadvance/specter/devices/device_types.py
@@ -1,0 +1,11 @@
+class DeviceTypes:
+    BITBOX02 = "bitbox02"
+    BITCOINCORE = "bitcoincore"
+    BITCOINCORE_WATCHONLY = "bitcoincore_watchonly"
+    COBO = "cobo"
+    COLDCARD = "coldcard"
+    ELECTRUM = "electrum"
+    GENERICDEVICE = "other"
+    KEEPKEY = "keepkey"
+    SPECTER = "specter"
+    TREZOR = "trezor"
